More Comprehensive Setup
Added loop to set up previous topics in a single script, that way there won't be any issues with dependencies.
Added a negated subclass check to Topic for the above loop.
Added a __coreSetup__ method to the Slide decorator class and a call to it that gets rid of the need for repeated code in the slide setup descriptors.